simplify test cases
