+++ tests/failpoints/cases/test_gc_worker.rs
@@ -6,6 +6,7 @@ use std::thread;
 use std::time::Duration;
 
 use collections::HashMap;
+use engine_rocks::RocksWriteBatch;
 use engine_traits::Peekable;
 use grpcio::{ChannelBuilder, Environment};
 use keys::data_key;
@@ -17,6 +18,9 @@ use test_raftstore::*;
 use tikv::server::gc_worker::{
     AutoGcConfig, GcSafePointProvider, Result as GcWorkerResult, TestGCRunner,
 };
+use tikv::storage::kv::TestEngineBuilder;
+use tikv::storage::mvcc::tests::must_get_none;
+use tikv::storage::txn::tests::{must_commit, must_prewrite_delete, must_prewrite_put};
 use tikv_util::HandyRwLock;
 use txn_types::{Key, TimeStamp};
 
@@ -290,6 +294,37 @@ fn test_collect_applying_locks() {
     // Must observe the applying lock even through we can't use scan to get it.
     assert_eq!(locks.len(), 1);
     assert_eq!(locks[0].get_key(), b"k1");
+}
+
+// Test write CF's compaction filter can call `orphan_versions_handler` correctly.
+#[test]
+fn test_error_in_compaction_filter() {
+    let engine = TestEngineBuilder::new().build().unwrap();
+    let raw_engine = engine.get_rocksdb();
+
+    let large_value = vec![b'x'; 300];
+    must_prewrite_put(&engine, b"zkey", &large_value, b"zkey", 101);
+    must_commit(&engine, b"zkey", 101, 102);
+    must_prewrite_put(&engine, b"zkey", &large_value, b"zkey", 103);
+    must_commit(&engine, b"zkey", 103, 104);
+    must_prewrite_delete(&engine, b"zkey", b"zkey", 105);
+    must_commit(&engine, b"zkey", 105, 106);
+
+    let fp = "write_compaction_filter_flush_write_batch";
+    fail::cfg(fp, "return").unwrap();
+
+    let mut gc_runner = TestGCRunner::default();
+    gc_runner.safe_point = 200;
+    gc_runner.orphan_versions_handler = Some(Arc::new(move |wb: RocksWriteBatch| {
+        assert_eq!(wb.as_inner().count(), 2);
+    }));
+    gc_runner.gc(&raw_engine);
+
+    // Although versions on default CF is not cleaned, write CF is GCed correctly.
+    must_get_none(&engine, b"zkey", 102);
+    must_get_none(&engine, b"zkey", 104);
+
+    fail::remove(fp);
 }
 
 // Test GC worker can receive and handle orphan versions emit from write CF's compaction filter
